+++ app/admin/job.rb
@@ -102,11 +102,11 @@ ActiveAdmin.register Job do
     render :new, layout: false
   end
 
-  action_item only: :show do
+  action_item :clone, only: :show do
     link_to(I18n.t('admin.job.clone'), clone_admin_job_path(id: job.id))
   end
 
-  sidebar :relations, only: :show do
+  sidebar :relations, only: [:show, :edit] do
     job_query = AdminHelpers::Link.query(:job_id, job.id)
 
     ul do
@@ -131,7 +131,7 @@ ActiveAdmin.register Job do
     end
   end
 
-  sidebar :app, only: :show do
+  sidebar :app, only: [:show, :edit] do
     ul do
       # rubocop:disable Metrics/LineLength
       li link_to I18n.t('admin.view_in_app.view'), FrontendRouter.draw(:job, id: job.id), target: '_blank'
@@ -140,7 +140,7 @@ ActiveAdmin.register Job do
     end
   end
 
-  sidebar :latest_applicants, only: :show do
+  sidebar :latest_applicants, only: [:show, :edit] do
     ul do
       job.job_users.
         order(created_at: :desc).