+++ app/models/user_skill.rb
@@ -1,5 +1,8 @@
 # frozen_string_literal: true
 class UserSkill < ApplicationRecord
+  PROFICIENCY_RANGE = 1..10
+  PROFICIENCY_ADMIN_RANGE = 1..7
+
   belongs_to :user
   belongs_to :skill
 
@@ -7,17 +10,38 @@ class UserSkill < ApplicationRecord
   validates :user, presence: true
   validates :skill, uniqueness: { scope: :user }
   validates :user, uniqueness: { scope: :skill }
+
+  scope :visible, -> { joins(:skill).where(skills: { internal: false }) }
+
+  def self.safe_create(skill:, user:, proficiency_by_admin: nil)
+    return find_or_create_by(user: user, skill: skill) if proficiency_by_admin.blank?
+
+    find_or_initialize_by(user: user, skill: skill).tap do |user_skill|
+      user_skill.proficiency_by_admin = proficiency_by_admin
+      user_skill.save
+    end
+  end
+
+  def self.safe_destroy(skill:, user:)
+    find_by(user: user, skill: skill)&.destroy!
+  end
+
+  def touched_by_admin?
+    !proficiency_by_admin.nil?
+  end
 end
 
 # == Schema Information
 #
 # Table name: user_skills
 #
-#  id         :integer          not null, primary key
-#  user_id    :integer
-#  skill_id   :integer
-#  created_at :datetime         not null
-#  updated_at :datetime         not null
+#  id                   :integer          not null, primary key
+#  user_id              :integer
+#  skill_id             :integer
+#  created_at           :datetime         not null
+#  updated_at           :datetime         not null
+#  proficiency          :integer
+#  proficiency_by_admin :integer
 #
 # Indexes
 #