+++ spec/models/user_skill_spec.rb
@@ -2,6 +2,56 @@
 require 'rails_helper'
 
 RSpec.describe UserSkill, type: :model do
+  let(:user) { FactoryGirl.create(:user) }
+  let(:skill) { FactoryGirl.create(:skill) }
+
+  describe '#touched_by_admin?' do
+    it 'returns true if proficiency_by_admin is set' do
+      user_skill = FactoryGirl.build(:user_skill, proficiency_by_admin: 7)
+      expect(user_skill.touched_by_admin?).to eq(true)
+    end
+
+    it 'returns false if proficiency_by_admin is not set' do
+      user_skill = FactoryGirl.build(:user_skill, proficiency_by_admin: nil)
+      expect(user_skill.touched_by_admin?).to eq(false)
+    end
+  end
+
+  describe '#safe_create' do
+    it 'can create user skill' do
+      expect do
+        user_skill = described_class.safe_create(skill: skill, user: user)
+
+        expect(user_skill).to be_persisted
+      end.to change(described_class, :count).by(1)
+    end
+
+    it 'can safely invoke create user skill twice' do
+      expect do
+        user_skill = described_class.safe_create(skill: skill, user: user)
+        described_class.safe_create(skill: skill, user: user)
+
+        expect(user_skill).to be_persisted
+      end.to change(described_class, :count).by(1)
+    end
+  end
+
+  describe '#safe_destroy' do
+    it 'can delete user skill' do
+      described_class.safe_create(skill: skill, user: user)
+      expect do
+        described_class.safe_destroy(skill: skill, user: user)
+      end.to change(described_class, :count).by(-1)
+    end
+
+    it 'can safely invoke delete user skill twice' do
+      described_class.safe_create(skill: skill, user: user)
+      expect do
+        described_class.safe_destroy(skill: skill, user: user)
+        described_class.safe_destroy(skill: skill, user: user)
+      end.to change(described_class, :count).by(-1)
+    end
+  end
 end
 
 # == Schema Information